refactor: put voice interactive button out of reaction controls to control bar
